prod/requirements.txt
apache-airflow
pandas
requests
google-cloud-storage